+++ lib/src/values/typedefs.dart
@@ -97,3 +97,7 @@ typedef ChatTextFieldViewBuilderCallback<T> = Widget Function(
   T value,
   Widget? child,
 );
+typedef TextFieldActionWidgetBuilder = List<Widget> Function(
+  BuildContext context,
+  TextEditingController controller,
+);